split chat playground component for better readability
