+++ src/components/openai-images-playground/image-gallery.tsx
