+++ src/components/openai-images-playground/image-promp-form.tsx
